Moved requireLoggedIn files
